+++ test/webpack.mock2.ts
@@ -5,7 +5,8 @@ export default webpackMockServer.add((app, helper) => {
     const response = [];
     for (let i = 0; i < 101; ++i) {
       response.push({
-        id: helper.getRandomInt(),
+        id: helper.getUniqueIdInt(),
+        randomInt: helper.getRandomInt(),
         lastDate: new Date()
       });
     }